Fix windows build.
git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24181 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/representation-change.h
@@ -126,10 +126,10 @@ class RepresentationChanger {
       case IrOpcode::kInt32Constant:
         if (output_type & kTypeUint32) {
           uint32_t value = OpParameter<uint32_t>(node);
-          return jsgraph()->Float32Constant(value);
+          return jsgraph()->Float32Constant(static_cast<float>(value));
         } else {
           int32_t value = OpParameter<int32_t>(node);
-          return jsgraph()->Float32Constant(value);
+          return jsgraph()->Float32Constant(static_cast<float>(value));
         }
       case IrOpcode::kFloat32Constant:
         return node;  // No change necessary.
